* fix build under x86_64 arch.
CCMAIL: kde-bindings@kde.org

svn path=/branches/KDE/4.2/kdebindings/falcon/krossfalcon/; revision=912343

+++ falcon/falconkfvm.cpp
@@ -1399,7 +1399,7 @@ static void init_reflect_object_func( Falcon::VMachine *vm )
     Falcon::ExtFuncDef *def = vm->currentSymbol()->getExtFuncDef();
     
     // The dynamic type manager has recorded the meta type here.
-    int metaType = (int) def->extra();
+    long metaType = (long) def->extra();
     if ( metaType == 0 )
     {
         vm->raiseModError( new Falcon::CodeError(
